Add files via upload
Finished off the voltage monitoring from the phone.

Needs a 1.5k and 470 Ohm voltage divider from the battery to A0 as it uses the internal reference to make it stable. Colour of the voltage monitor turns from green to red when it gets to below 3.7v. as a reminder to charge it.